Renamed Chart component to Charts, and charts folder to chart
