+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as Filters } from "./Filters"
-export { default as Chart } from "./Chart"
+export { default as Charts } from "./Charts"
 export { default as Papers } from "./Papers"
 export { default as TagList } from "./TagList"
 export { default as AddPaperForm } from "./Form"
